+++ testRoundTrip/examples/RoehFlux_MemOps/driver_scalar_dd.f90
@@ -72,7 +72,7 @@ program driver
   double precision flux_ph1, flux_ph2, flux_ph3, flux_ph4, flux_ph5
 
   double precision, dimension(19) :: x0
-  real :: h
+  double precision :: h
   integer:: n,m
   integer i,j,k
